# Parcial3/Swagger/swagger-definition.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Usuarios",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "description": "Obtiene todos los usuarios.",
        "responses": {
          "200": {
            "description": "Se obtienen los usuarios con éxito."
          }
        }
      },
      "post": {
        "description": "Crea un nuevo usuario.",
        "parameters": [
          {
            "in": "query",
            "name": "nombre",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "El nombre del usuario."
          },
          {
            "in": "query",
            "name": "edad",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "La edad del usuario."
          },
          {
            "in": "query",
            "name": "juegoFavorito",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "El juego favorito del usuario."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario creado con éxito."
          },
          "400": {
            "description": "Error al agregar la información o campos vacíos/nulos."
          }
        }
      },
      "delete": {
        "description": "Elimina un usuario por ID.",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "El ID del usuario a eliminar."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado con éxito."
          },
          "404": {
            "description": "No se encontró el usuario para eliminar."
          },
          "400": {
            "description": "Error al eliminar el usuario o ID vacío/nulo."
          }
        }
      },
      "put": {
        "description": "Actualiza un usuario por ID.",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "El ID del usuario a actualizar."
          },
          {
            "in": "query",
            "name": "nombre",
            "schema": {
              "type": "string"
            },
            "description": "Nuevo nombre del usuario."
          },
          {
            "in": "query",
            "name": "edad",
            "schema": {
              "type": "integer"
            },
            "description": "Nueva edad del usuario."
          },
          {
            "in": "query",
            "name": "juegoFavorito",
            "schema": {
              "type": "string"
            },
            "description": "Nuevo juego favorito del usuario."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario actualizado con éxito."
          },
          "404": {
            "description": "No se encontró el usuario para actualizar."
          },
          "400": {
            "description": "Error al actualizar el usuario o ID vacío/nulo."
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}